# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TCFL Portal API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/apply/application/programme": {
      "put": {
        "tags": [
          "apply"
        ],
        "summary": "Choose Programme",
        "description": "Starts the application on first choice; the programme can change until it's submitted.",
        "operationId": "choose_programme",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgrammeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyApplication"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/apply/offer": {
      "post": {
        "tags": [
          "apply"
        ],
        "summary": "Answer Offer",
        "operationId": "answer_offer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyApplication"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "draft",
          "submitted",
          "in_review",
          "more_info",
          "accepted",
          "rejected",
          "withdrawn"
        ],
        "title": "ApplicationStatus"
      },
      "ApplyStep": {
        "type": "string",
        "enum": [
          "programme",
          "national_id",
          "birth_certificate",
          "results",
          "review",
          "submit"
        ],
        "title": "ApplyStep"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InfoRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "message",
          "at"
        ],
        "title": "InfoRequest"
      },
...
      "MyApplication": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "programme_id": {
            "type": "string",
            "format": "uuid",
            "title": "Programme Id"
          },
          "programme": {
            "type": "string",
            "title": "Programme"
          },
          "intake": {
            "type": "string",
            "title": "Intake"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "initials": {
            "type": "string",
            "title": "Initials"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "review_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Started At"
          },
          "decision_expected_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Expected By"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "requests": {
            "items": {
              "$ref": "#/components/schemas/InfoRequest"
            },
            "type": "array",
            "title": "Requests"
          },
          "steps": {
            "$ref": "#/components/schemas/Steps"
          },
          "next_step": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplyStep"
              },
              {
                "type": "null"
              }
            ]
          },
          "phone_masked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Masked"
          },
          "offer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Offer"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "reference",
          "status",
          "programme_id",
          "programme",
          "intake",
          "name",
          "initials",
          "created_at",
          "submitted_at",
          "review_started_at",
          "decision_expected_by",
          "decided_at",
          "decision_reason",
          "requests",
          "steps",
          "next_step",
          "phone_masked",
          "offer"
        ],
        "title": "MyApplication"
      },
...
      "Offer": {
        "properties": {
          "starts_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          },
          "accept_by": {
            "type": "string",
            "format": "date",
            "title": "Accept By"
          },
          "registration_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration At"
          },
          "registration_place": {
            "type": "string",
            "title": "Registration Place"
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "declined_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declined At"
          }
        },
        "type": "object",
        "required": [
          "starts_on",
          "accept_by",
          "registration_at",
          "registration_place",
          "accepted_at",
          "declined_at"
        ],
        "title": "Offer"
      },
      "OfferAnswer": {
        "type": "string",
        "enum": [
          "accept",
          "decline"
        ],
        "title": "OfferAnswer"
      },
      "OfferIn": {
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/OfferAnswer"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "OfferIn"
      },
...
      "ProgrammeIn": {
        "properties": {
          "programme_id": {
            "type": "string",
            "format": "uuid",
            "title": "Programme Id"
          }
        },
        "type": "object",
        "required": [
          "programme_id"
        ],
        "title": "ProgrammeIn"
      },
...
      "Steps": {
        "properties": {
          "programme": {
            "type": "boolean",
            "title": "Programme"
          },
          "national_id": {
            "type": "boolean",
            "title": "National Id"
          },
          "birth_certificate": {
            "type": "boolean",
            "title": "Birth Certificate"
          },
          "results": {
            "type": "boolean",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "programme",
          "national_id",
          "birth_certificate",
          "results"
        ],
        "title": "Steps"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}